requirements.txt
astroid==2.1.0
atomicwrites==1.3.0
attrs==18.2.0
autoenv==1.0.0
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
coverage==4.5.2
coveralls==1.5.1
docopt==0.6.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT==0.3.2
Flask-JWT-Extended==3.17.0
Flask-Mail==0.9.1
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==5.0.0
pep8==1.7.1
pluggy==0.8.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
py==1.7.0
PyJWT==1.4.2
pylint==2.2.2
pytest==4.2.0
pytest-cov==2.6.1
python-dotenv==0.10.1
requests==2.21.0
six==1.12.0
typed-ast==1.3.0
urllib3==1.24.1
virtualenv==16.4.0
Werkzeug==0.14.1
wrapt==1.11.1
